envoy: Move test package from pkg/proxy to pkg/envoy
Move test package from pkg/proxy to pkg/envoy as that is the only place
it is used from.
